requirements.txt
requests
python-dotenv
numpy
pyaudio
pygame
mss
Pillow
whisper
spotipy
pyperclip
playwright
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dateutil
tiktoken